+++ src/audio/copier/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_local_sources(sof copier.c copier_hifi.c copier_generic.c)
+
